milestone_1/requirements.txt
pandas
numpy
scikit-learn
typing
PyYAML
requests
sqlalchemy
psycopg2-binary